+++ builder/src/feat/mod.rs
@@ -1,6 +1,7 @@
 //! Feats that a character can have.
 public_modules!(feats, requirements);
 
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use utils::public_modules;
 
@@ -43,7 +44,7 @@ impl CloneBonus for Feat {
 }
 
 impl GetBonuses for Feat {
-    fn get_bonuses(&self, value: f32) -> Option<Vec<crate::bonus::Bonus>> {
+    fn get_bonuses(&self, value: Decimal) -> Option<Vec<crate::bonus::Bonus>> {
         match self {
             Self::RacialFeat(feat) => feat.get_bonuses(value),
             Self::Proficiency(_) => None,
